+++ src/features/resurrection/hooks/useResurrection.ts
@@ -72,7 +72,7 @@ export function useResurrection(sarcoId: string, recipientPrivateKey: string) {
     } catch (error) {
       throw new Error(`Error resurrecting sarcophagus: ${error}`);
     }
-  }, [archaeologists, canResurrect, recipientPrivateKey, sarcoId, sarcophagus]);
+  }, [archaeologists, canResurrect, recipientPrivateKey, sarcoId, sarcophagus, fetchArweaveFileFallback]);
 
   return { canResurrect, resurrect };
 }
